feat(trustscan): refactor TrustScan service with new scoring, confidence, and summary features; modularize factor building and constants
