+++ server/src/controllers/trustScanController.js
@@ -3,8 +3,16 @@ import { validationResult } from "express-validator";
 import TrustScanJob from "../models/TrustScanJob.js";
 import TrustScanReport from "../models/TrustScanReport.js";
 import { checkDomainSignals, checkSecurityHeaders, runSslTlsCheck } from "../services/trustScanSignals.js";
-
-const MOCK_SCAN_DURATION_MS = 4500;
+import { MOCK_SCAN_DURATION_MS } from "../services/trustscan/constants.js";
+import { getTrustScanConfidence } from "../services/trustscan/confidenceService.js";
+import {
+  buildDomainFactor,
+  buildHeadersFactor,
+  buildSslFactor,
+  createPlaceholderFactors
+} from "../services/trustscan/factorBuilders.js";
+import { calculateScoreAndVerdict } from "../services/trustscan/scoringService.js";
+import { buildTrustScanSummary } from "../services/trustscan/summaryService.js";
 
 const getNormalizedDomain = (rawUrl) => {
   try {
@@ -24,146 +32,6 @@ const getNormalizedDomain = (rawUrl) => {
   }
 };
 
-const createPlaceholderFactors = () => [
-  {
-    key: "headers",
-    label: "Security Headers",
-    impact: 0,
-    status: "pending",
-    detail: "Security headers signal is queued for implementation."
-  },
-  {
-    key: "reputation",
-    label: "Reputation",
-    impact: 0,
-    status: "pending",
-    detail: "Reputation signal is queued for implementation."
-  }
-];
-
-const getVerdictBand = (score) => {
-  if (score <= 30) return "DANGEROUS";
-  if (score <= 55) return "RISKY";
-  if (score <= 75) return "CAUTION";
-  if (score <= 89) return "SAFE";
-  return "STRONG";
-};
-
-const buildSslFactor = (ssl) => {
-  if (ssl.valid) {
-    const expiryText = typeof ssl.daysRemaining === "number"
-      ? `${ssl.daysRemaining} days remaining`
-      : "expiry unknown";
-
-    return {
-      key: "ssl",
-      label: "Transport Security",
-      impact: 15,
-      status: "pass",
-      detail: `Valid certificate from ${ssl.issuer}. Expires at ${ssl.expiresAt || "unknown"} (${expiryText}).`
-    };
-  }
-
-  return {
-    key: "ssl",
-    label: "Transport Security",
-    impact: -35,
-    status: "fail",
-    detail: `Certificate check failed: ${ssl.error || "invalid or expired certificate"}. Issuer: ${ssl.issuer}.`
-  };
-};
-
-const buildHeadersFactor = (headers) => {
-  const isStrong = headers.grade === "Strong";
-  const isGood = headers.grade === "Good";
-
-  return {
-    key: "headers",
-    label: "Browser Protection Controls",
-    impact: headers.scoreDelta,
-    status: isStrong || isGood ? "pass" : headers.missing.length <= 3 ? "warn" : "fail",
-    detail: `${headers.grade} browser protection posture. Present: ${headers.present.join(", ") || "none"}. Missing: ${headers.missing.join(", ") || "none"}.`,
-    present: headers.present,
-    missing: headers.missing,
-    grade: headers.grade
-  };
-};
-
-const buildDomainFactor = (domain) => {
-  const ageText = typeof domain.ageDays === "number"
-    ? `${domain.ageDays} days old`
-    : "domain age unavailable";
-  const resolvesText = domain.resolves ? "DNS resolves cleanly" : "DNS resolution failed";
-  const mxText = domain.mx ? "MX records present" : "No MX records detected";
-  const nameserverText = domain.nameservers > 0
-    ? `${domain.nameservers} nameserver${domain.nameservers === 1 ? "" : "s"}`
-    : "No nameservers detected";
-
-  let status = "pass";
-  if (!domain.resolves) {
-    status = "fail";
-  } else if (domain.grade === "Suspicious" || domain.scoreDelta < -15) {
-    status = "warn";
-  }
-
-  return {
-    key: "dns",
-    label: "Domain Intelligence",
-    impact: domain.scoreDelta,
-    status,
-    detail: `${resolvesText}. ${mxText}. ${nameserverText}. ${ageText}.`,
-    resolves: domain.resolves,
-    mx: domain.mx,
-    ageDays: domain.ageDays,
-    nameservers: domain.nameservers,
-    grade: domain.grade,
-    checkedDomain: domain.checkedDomain
-  };
-};
-
-const buildSummary = ({ ssl, headers, domain }) => {
-  const tlsClause = ssl.valid
-    ? `valid TLS from ${ssl.issuer}`
-    : `TLS validation failed (${ssl.error || "invalid or expired certificate"})`;
-
-  const headerClause = headers.missing.length === 0
-    ? "strong browser protection controls"
-    : `${headers.missing.slice(0, 3).join(", ")} browser protections missing`;
-
-  let domainClause = "domain intelligence is inconclusive";
-
-  if (!domain.resolves) {
-    domainClause = "the domain fails DNS resolution";
-  } else if (typeof domain.ageDays === "number" && domain.ageDays < 30) {
-    domainClause = `a newly registered domain (${domain.ageDays} days old)`;
-  } else if (typeof domain.ageDays === "number" && domain.ageDays < 180) {
-    domainClause = `a relatively new domain (${domain.ageDays} days old)`;
-  } else if (typeof domain.ageDays === "number") {
-    domainClause = `an established domain (${domain.ageDays} days old)`;
-  } else {
-    domainClause = "partial domain intelligence coverage";
-  }
-
-  return `Site has ${tlsClause}, ${headerClause}, and ${domainClause}.`;
-};
-
-const buildConfidenceLabel = ({ ssl, headers, domain }) => {
-  const sslReady = Boolean(ssl) && (ssl.issuer !== "Unknown" || ssl.expiresAt !== null || typeof ssl.daysRemaining === "number");
-  const headersReady = Boolean(headers) && Boolean(headers.checkedUrl);
-  const domainReady = Boolean(domain) && Boolean(domain.checkedDomain);
-  const ageReady = typeof domain?.ageDays === "number";
-
-  if (sslReady && headersReady && domainReady && ageReady) {
-    return "High";
-  }
-
-  if ((sslReady && headersReady && domainReady) || ((sslReady || headersReady) && domainReady)) {
-    return "Medium";
-  }
-
-  return "Low";
-};
-
 const buildMockReportPayload = async (job) => {
   const [ssl, headers, domain] = await Promise.all([
     runSslTlsCheck(job.url),
@@ -177,9 +45,7 @@ const buildMockReportPayload = async (job) => {
     buildDomainFactor(domain),
     ...createPlaceholderFactors()
   ];
-
-  const baseScore = 70;
-  const score = Math.max(0, Math.min(100, baseScore + factors.reduce((sum, item) => sum + item.impact, 0)));
+  const { score, verdict } = calculateScoreAndVerdict(factors);
 
   return {
     jobId: job._id,
@@ -187,10 +53,10 @@ const buildMockReportPayload = async (job) => {
     url: job.url,
     normalizedDomain: job.normalizedDomain,
     score,
-    verdict: getVerdictBand(score),
-    confidence: buildConfidenceLabel({ ssl, headers, domain }),
+    verdict,
+    confidence: getTrustScanConfidence({ ssl, headers, domain }),
     factors,
-    summary: buildSummary({ ssl, headers, domain })
+    summary: buildTrustScanSummary({ ssl, headers, domain })
   };
 };
