+++ server/src/services/trustscan/index.js
@@ -1,3 +1,13 @@
 export { runSslTlsCheck } from "./sslSignal.js";
 export { checkSecurityHeaders } from "./headerSignal.js";
 export { checkDomainSignals } from "./domainSignal.js";
+export { calculateScoreAndVerdict } from "./scoringService.js";
+export { buildTrustScanSummary } from "./summaryService.js";
+export { getTrustScanConfidence } from "./confidenceService.js";
+export {
+	buildDomainFactor,
+	buildHeadersFactor,
+	buildSslFactor,
+	createPlaceholderFactors
+} from "./factorBuilders.js";
+export { BASE_SCORE, MOCK_SCAN_DURATION_MS, VERDICT_BANDS } from "./constants.js";
